+++ library/Mockery/Adapter/Phpunit/TestListener.php
@@ -14,86 +14,51 @@
  *
  * @category   Mockery
  * @package    Mockery
- * @copyright  Copyright (c) 2010-2014 Pádraic Brady (http://blog.astrumfutura.com)
+ * @copyright  Copyright (c) 2010 Pádraic Brady (http://blog.astrumfutura.com)
  * @license    http://github.com/padraic/mockery/blob/master/LICENSE New BSD License
  */
 
 namespace Mockery\Adapter\Phpunit;
 
-class TestListener implements \PHPUnit_Framework_TestListener
+class TestListener extends \PHPUnit_Framework_BaseTestListener
 {
     /**
-     * After each test, perform Mockery verification tasks and cleanup the
-     * statically stored Mockery container for the next test.
+     * endTest is called after each test and checks if \Mockery::close() has
+     * been called, and will let the test fail if it hasn't.
      *
      * @param  PHPUnit_Framework_Test $test
      * @param  float                  $time
      */
     public function endTest(\PHPUnit_Framework_Test $test, $time)
     {
+        if (!$test instanceof \PHPUnit_Framework_TestCase) {
+            // We need the getTestResultObject and getStatus methods which are
+            // not part of the interface.
+            return;
+        }
+
+        if ($test->getStatus() !== \PHPUnit_Runner_BaseTestRunner::STATUS_PASSED) {
+            // If the test didn't pass there is no guarantee that
+            // verifyMockObjects and assertPostConditions have been called.
+            // And even if it did, the point here is to prevent false
+            // negatives, not to make failing tests fail for more reasons.
+            return;
+        }
+
         try {
-            $container = \Mockery::getContainer();
-            // check addToAssertionCount is important to avoid mask test errors
-            if ($container != null && method_exists($test, 'addToAssertionCount')) {
-                $expectation_count = $container->mockery_getExpectationCount();
-                $test->addToAssertionCount($expectation_count);
-            }
-            \Mockery::close();
-        } catch (\Exception $e) {
-            $result = $test->getTestResultObject();
-            $result->addError($test, $e, $time);
+            // The self() call is used as a sentinel. Anything that throws if
+            // the container is closed already will do.
+            \Mockery::self();
+        } catch (\LogicException $_) {
+            return;
         }
-    }
 
-    /**
-     * Add Mockery files to PHPUnit's blacklist so they don't showup on coverage reports
-     */
-    public function startTestSuite(\PHPUnit_Framework_TestSuite $suite)
-    {
-        if (class_exists('\\PHP_CodeCoverage_Filter')
-        && method_exists('\\PHP_CodeCoverage_Filter', 'getInstance')) {
-            \PHP_CodeCoverage_Filter::getInstance()->addDirectoryToBlacklist(
-                __DIR__ . '/../../../Mockery/',
-                '.php',
-                '',
-                'PHPUNIT'
-            );
-
-            \PHP_CodeCoverage_Filter::getInstance()->addFileToBlacklist(__DIR__.'/../../../Mockery.php', 'PHPUNIT');
-        }
-    }
-    /**
-     *  The Listening methods below are not required for Mockery
-     */
-    public function addError(\PHPUnit_Framework_Test $test, \Exception $e, $time)
-    {
-    }
-
-    public function addWarning(\PHPUnit_Framework_Test $test, \PHPUnit_Framework_Warning $e, $time)
-    {
-    }
-
-    public function addFailure(\PHPUnit_Framework_Test $test, \PHPUnit_Framework_AssertionFailedError $e, $time)
-    {
-    }
-
-    public function addIncompleteTest(\PHPUnit_Framework_Test $test, \Exception $e, $time)
-    {
-    }
-
-    public function addSkippedTest(\PHPUnit_Framework_Test $test, \Exception $e, $time)
-    {
-    }
-
-    public function addRiskyTest(\PHPUnit_Framework_Test $test, \Exception $e, $time)
-    {
-    }
-
-    public function endTestSuite(\PHPUnit_Framework_TestSuite $suite)
-    {
-    }
-
-    public function startTest(\PHPUnit_Framework_Test $test)
-    {
+        $e = new \PHPUnit_Framework_ExpectationFailedException(sprintf(
+            "Mockery's expectations have not been verified. Make sure that \Mockery::close() is called at the end of the test. Consider using %s\MockeryPHPUnitIntegration or extending %s\MockeryTestCase.",
+            __NAMESPACE__,
+            __NAMESPACE__
+        ));
+        $result = $test->getTestResultObject();
+        $result->addFailure($test, $e, $time);
     }
 }