+++ library/Mockery/Generator/MockConfiguration.php
@@ -1,4 +1,22 @@
 <?php
+/**
+ * Mockery
+ *
+ * LICENSE
+ *
+ * This source file is subject to the new BSD license that is bundled
+ * with this package in the file LICENSE.txt.
+ * It is also available through the world-wide-web at this URL:
+ * http://github.com/padraic/mockery/blob/master/LICENSE
+ * If you did not receive a copy of the license and are unable to
+ * obtain it through the world-wide-web, please send an email
+ * to padraic@php.net so we can send you a copy immediately.
+ *
+ * @category   Mockery
+ * @package    Mockery
+ * @copyright  Copyright (c) 2010 Pádraic Brady (http://blog.astrumfutura.com)
+ * @license    http://github.com/padraic/mockery/blob/master/LICENSE New BSD License
+ */
 
 namespace Mockery\Generator;
 
@@ -294,7 +312,7 @@ class MockConfiguration
 
             $this->targetClass = $dtc;
         } else {
-            $this->targetClass = new UndefinedTargetClass($this->targetClassName);
+            $this->targetClass = UndefinedTargetClass::factory($this->targetClassName);
         }
 
         return $this->targetClass;
@@ -308,7 +326,7 @@ class MockConfiguration
 
         foreach ($this->targetInterfaceNames as $targetInterface) {
             if (!interface_exists($targetInterface)) {
-                $this->targetInterfaces[] = new UndefinedTargetClass($targetInterface);
+                $this->targetInterfaces[] = UndefinedTargetClass::factory($targetInterface);
                 return;
             }