+++ library/Mockery/Generator/StringManipulation/Pass/MethodDefinitionPass.php
@@ -1,4 +1,22 @@
 <?php
+/**
+ * Mockery
+ *
+ * LICENSE
+ *
+ * This source file is subject to the new BSD license that is bundled
+ * with this package in the file LICENSE.txt.
+ * It is also available through the world-wide-web at this URL:
+ * http://github.com/padraic/mockery/blob/master/LICENSE
+ * If you did not receive a copy of the license and are unable to
+ * obtain it through the world-wide-web, please send an email
+ * to padraic@php.net so we can send you a copy immediately.
+ *
+ * @category   Mockery
+ * @package    Mockery
+ * @copyright  Copyright (c) 2010 Pádraic Brady (http://blog.astrumfutura.com)
+ * @license    http://github.com/padraic/mockery/blob/master/LICENSE New BSD License
+ */
 
 namespace Mockery\Generator\StringManipulation\Pass;
 
@@ -91,12 +109,20 @@ class MethodDefinitionPass implements Pass
             'bool',
             'float',
             'int',
-            'string'
+            'string',
+            'iterable',
         );
         $typeHint = trim($param->getTypeHintAsString());
 
-        if (!empty($typeHint) && !in_array($typeHint, $languageTypeHints)) {
-            $typeHint = '\\'.$typeHint;
+        if (!empty($typeHint)) {
+           
+            if (!in_array($typeHint, $languageTypeHints)) {
+                $typeHint = '\\'.$typeHint;
+            }
+
+            if (version_compare(PHP_VERSION, '7.1.0-dev') >= 0 && $param->allowsNull()) {
+                $typeHint = "?".$typeHint;
+            }
         }
 
         return $typeHint .= ' ';
@@ -148,11 +174,14 @@ if (\$argc > $i) {
 BODY;
             }
         }
-        $body .= <<<BODY
-\$ret = {$invoke}(__FUNCTION__, \$argv);
-return \$ret;
-}
-BODY;
+
+        $body .= "\$ret = {$invoke}(__FUNCTION__, \$argv);\n";
+
+        if ($method->getReturnType() !== "void") {
+            $body .= "return \$ret;\n";
+        }
+
+        $body .= "}\n";
         return $body;
     }
 }